feat: criação do id de cada usuário

+++ src/app/cadastro/page.tsx
@@ -4,10 +4,18 @@ import Button from "@/components/Button";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 
+type User = {
+  id: number;
+  name: string;
+  email: string;
+  password: string;
+};
+
 export default function SignIn() {
   const [name, setName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [error, setError] = useState("");
 
   const getUsers = () => {
     const storedUsers = localStorage.getItem("@Users");
@@ -21,12 +29,18 @@ export default function SignIn() {
     if (name.trim() !== "" && email.trim() !== "" && password.trim() !== "") {
       const users = getUsers();
 
-      const newUser = { name, email, password };
+      const emailExists = users.some((item: User) => item.email === email);
 
+      if (emailExists) {
+        setError("Esse email já foi cadastrado. Faça o logIn");
+        return;
+      }
+
+      setError("");
+      const id = users.length + 1;
+      const newUser = { id, name, email, password };
       users.push(newUser);
-
       localStorage.setItem("@Users", JSON.stringify(users));
-
       router.push("/login");
     }
   };
@@ -34,9 +48,15 @@ export default function SignIn() {
   return (
     <div className="flex flex-col h-screen w-screen justify-center items-center">
       <form
-        className="flex flex-col border border-white/30 w-96 h-96 rounded-md p-4 gap-3 justify-center items-center "
+        className="flex flex-col border border-white/30 w-96 h-auto rounded-md p-4 gap-3 justify-center items-center "
         id="signInId"
       >
+        {error && (
+          <div className="bg-red-400 text-black border border-red-800 rounded-md p-3 mb-4">
+            {error}
+          </div>
+        )}
+
         <h1 className="font-bold text-xl mb-3">Faça seu cadastro!</h1>
         <input
           value={name}
@@ -60,6 +80,14 @@ export default function SignIn() {
           onChange={(e) => setPassword(e.target.value)}
         />
         <Button type="button" onClick={addUser} />
+
+        <p className=" mt-4 text-sm" onClick={() => router.push("/cadastro")}>
+          Já tem uma conta? Então faça o{" "}
+          <span className="cursor-pointer hover:text-blue-700 duration-200">
+            LogIn
+          </span>{" "}
+          agora mesmo!
+        </p>
       </form>
     </div>
   );
